Refactor snapshot testing
