+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewBroadcast@3x.png
