+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewButton@3x.png
