+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewConsentTransparencyInfo@3x.png
