+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewInlineConsent@3x.png
