+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewLabel@3x.png
