+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewRibbon@3x.png
