+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewRoundedImageView@3x.png
