+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewSwitchView@3x.png
