+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewTextField@3x.png
