+++ SnapshotTests/ReferenceImages_64/UnitTests.ComponentViewTests/testComponentViewToast@3x.png
