+++ SnapshotTests/ReferenceImages_64/UnitTests.FullscreenViewTests/testFullscreenViewEmptyView@3x.png
