+++ SnapshotTests/ReferenceImages_64/UnitTests.FullscreenViewTests/testFullscreenViewFrontpageView@3x.png
