+++ SnapshotTests/ReferenceImages_64/UnitTests.FullscreenViewTests/testFullscreenViewPopupView@3x.png
