+++ SnapshotTests/ReferenceImages_64/UnitTests.FullscreenViewTests/testFullscreenViewReviewView@3x.png
